pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "unihyperlora-experiments"
version = "0.1.0"
description = "Universal Hyper-LoRA experiments managed by Hydra"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1.0",
  "transformers>=4.36.0",
  "datasets>=2.14.0",
  "hydra-core>=1.3.2",
  "wandb>=0.16.0",
  "optuna>=3.4.0",
  "evaluate>=0.4.1",
  "seqeval>=1.2.2",
  "scikit-learn>=1.3.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "tqdm>=4.66.0",
  "pandas>=2.1.0",
]